Reorganize Screening-related spec files
[#142272305]
