+++ spec/javascripts/components/screenings/allegations/AllegationRowSpec.jsx
