+++ spec/javascripts/components/screenings/allegations/AllegationsCardViewSpec.jsx
