+++ spec/javascripts/components/screenings/allegations/AllegationsEditViewSpec.jsx
