+++ spec/javascripts/components/screenings/allegations/AllegationsShowViewSpec.jsx
