+++ spec/javascripts/components/screenings/crossReports/CrossReportCardViewSpec.jsx
