+++ spec/javascripts/components/screenings/decision/DecisionCardViewSpec.jsx
