+++ spec/javascripts/components/screenings/decision/DecisionShowViewSpec.jsx
