+++ spec/javascripts/components/screenings/incidentInformation/IncidentInformationCardViewSpec.jsx
