+++ spec/javascripts/components/screenings/incidentInformation/IncidentInformationShowViewSpec.jsx
