+++ spec/javascripts/components/screenings/narrative/NarrativeCardViewSpec.jsx
