+++ spec/javascripts/components/screenings/narrative/NarrativeEditViewSpec.jsx
