+++ spec/javascripts/components/screenings/narrative/NarrativeShowViewSpec.jsx
