+++ spec/javascripts/components/screenings/participants/ParticipantCardViewSpec.jsx
