+++ spec/javascripts/components/screenings/participants/ParticipantShowViewSpec.jsx
