+++ spec/javascripts/components/screenings/screeningInformation/ScreeningInformationCardViewSpec.jsx
