+++ spec/javascripts/components/screenings/screeningInformation/ScreeningInformationShowViewSpec.jsx
